+++ src/main/java/br/com/codex/v1/domain/dto/MotivoAcertoDto.java
@@ -11,7 +11,7 @@ public class MotivoAcertoDto implements Serializable {
     @Serial
     private static final long serialVersionUID = 1L;
 
-    protected Integer id;
+    protected Long id;
     protected String codigo;
     protected String descricao;
 
@@ -25,11 +25,11 @@ public class MotivoAcertoDto implements Serializable {
         this.descricao = obj.getDescricao();
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
