+++ src/main/java/br/com/codex/v1/domain/repository/AsoRepository.java
@@ -3,6 +3,6 @@ package br.com.codex.v1.domain.repository;
 import br.com.codex.v1.domain.rh.Aso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AsoRepository extends JpaRepository<Aso, Integer> {
+public interface AsoRepository extends JpaRepository<Aso, Long> {
 
 }
